requirements.txt
asn1crypto==0.24.0
beautifulsoup4==4.6.3
certifi==2018.10.15
cffi==1.11.5
chardet==3.0.4
cryptography==2.3.1
Django==1.11
django-redis==4.10.0
djangorestframework==3.9.0
idna==2.7
lxml==4.2.5
Pillow==5.3.0
pycparser==2.19
PyMySQL==0.9.2
pytz==2018.7
redis==3.0.1
requests==2.20.0
six==1.11.0
urllib3==1.24
